requirements.txt
amqp==5.2.0
ansible==9.2.0
ansible-core==2.16.3
ansible-runner==2.3.4
asgiref==3.7.2
async-timeout==4.0.3
bcrypt==4.1.2
billiard==4.2.0
celery==5.3.6
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.2
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docutils==0.20.1
gunicorn==21.2.0
Jinja2==3.1.3
jmespath==1.0.1
kombu==5.3.5
lockfile==0.12.2
MarkupSafe==2.1.5
numerize==0.12
packaging==23.2
paramiko==3.4.0
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.8.0
PyNaCl==1.5.0
python-daemon==3.0.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
resolvelib==1.0.1
secure==0.3.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
